Move when_any from stdexec to exec
